Remove effects on installation
The `postinstall` hook provides a minor convenience for developers
cloning the repo and developing within it, but only as far as
running the build for you whenever you install. The drawback to
using `postinstall` is that compilation runs whenever this package
is installed into another repo, which is totally unnecessary and
often (for JS users) results in failures when tsc or types packages
aren't installed within the project.

See also:
https://github.com/apollographql/typescript-repo-template/pull/84
https://github.com/apollographql/typescript-repo-template/pull/92

Fixes #40